Revert "🧹 Refactor init.py to use a model registry (#12138)"
This reverts commit f93326a214f4de914f5ffbbf8ab2a7bf0a2ea1de.
